Extracted duplicate localization tests logic into shared `XCTestCase` protocol
